requirements.txt
openai>=1.0.0
sqlalchemy>=2.0.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
numpy>=1.24.0
python-dotenv>=1.0.0
qdrant-client